+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Complete shell input with uppercase letters (e.g., `Cap` to `CapitalizedName` will now work)
+- Display UUID of created template in `storage templatise` output.
 
 ## [1.4.0] - 2022-06-15
 ### Added