+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0] - 2022-07-05
 ### Added
 - Add `--show-ip-addresses` flag to `server list` command to optionally include IP addresses in command output.
 - Add `database connection list`, `database connection cancel`, `database start`, and `database stop` commands.
@@ -139,3 +141,4 @@ Initial public beta release :tada:
 [1.2.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.2.0
 [1.3.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.3.0
 [1.4.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.4.0
+[1.5.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.5.0